LPS-63060 rename to portal-tools-data-partitioning-sql-builder
