+++ modules/util/portal-tools-data-partitioning-sql-builder/bnd.bnd
@@ -1,0 +1,9 @@
+Bundle-Name: Liferay Portal Tools Data Paritioning SQL Builder
+Bundle-SymbolicName: com.liferay.portal.tools.data.partitioning.sql.builder
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.tools.data.partitioning.sql.builder.exporter
+Main-Class: com.liferay.portal.tools.data.partitioning.sql.builder.Main
+-liferay-includeresource:\
+	@HikariCP-*.jar,\
+	@jcommander-*.jar,\
+	@slf4j-api-*.jar
