+++ modules/util/portal-tools-data-partitioning-sql-builder/build.gradle
