+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/resources/sample-mysql.properties
