--- modules/util/portal-tools-shard-builder/bnd.bnd
@@ -1,9 +1,0 @@
-Bundle-Name: Liferay Portal Tools Shard Builder
-Bundle-SymbolicName: com.liferay.portal.tools.shard.builder
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.portal.tools.shard.builder.exporter
-Main-Class: com.liferay.portal.tools.shard.builder.Main
--liferay-includeresource:\
-	@HikariCP-*.jar,\
-	@jcommander-*.jar,\
-	@slf4j-api-*.jar
