+++ source-formatter.properties
@@ -34,8 +34,8 @@
     # for finalable, mutable, or staticable field types.
     #
     check.java.field.types.excludes=\
+        modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/MainParameters.java,\
         modules/util/portal-tools-service-builder/src/main/java/com/liferay/portal/tools/service/builder/ServiceBuilder.java,\
-        modules/util/portal-tools-shard-builder/src/main/java/com/liferay/portal/tools/shard/builder/MainParameters.java,\
         modules/util/sync-engine/src/main/java/com/liferay/sync/engine/session/Session.java,\
         portal-impl/test/unit/com/liferay/portal/json/One.java,\
         portal-impl/test/unit/com/liferay/portal/json/Two.java,\
@@ -207,7 +207,7 @@
         **/markdown-converter/**,\
         **/osgi-util/**,\
         **/portal-osgi-web-wab-extender/**,\
-        **/portal-tools-shard-builder/**,\
+        **/portal-tools-data-partitioning-sql-builder/**,\
         **/registry-api/**,\
         **/sync-engine/**,\
         modules/test/poshi-runner/src/main/java/com/liferay/poshi/runner/util/GetterUtil.java,\
